Fix use of obsolete APIs in ExplainDataset and split classes out into separate files.
