+++ Libraries/dotNetRDF/Query/ExplainQueryProcessor.cs
@@ -39,167 +39,6 @@ using VDS.RDF.Writing.Formatting;
 namespace VDS.RDF.Query
 {
     /// <summary>
-    /// Represents the level of Query Explanation that is desired.
-    /// </summary>
-    [Flags]
-    public enum ExplanationLevel
-    {
-        /// <summary>
-        /// Specifies No Explanations
-        /// </summary>
-        None = 0,
-
-        /// <summary>
-        /// Specifies Explanations are output to Debug
-        /// </summary>
-        OutputToDebug = 1,
-
-        /// <summary>
-        /// Specifies Explanations are output to Trace
-        /// </summary>
-        OutputToTrace = 2,
-
-        /// <summary>
-        /// Specifies Explanations are output to Console Standard Output
-        /// </summary>
-        OutputToConsoleStdOut = 4,
-
-        /// <summary>
-        /// Specifies Explanations are output to Console Standard Error
-        /// </summary>
-        OutputToConsoleStdErr = 8,
-
-        /// <summary>
-        /// Specifies Explanations are output to Debug and Console Standard Output
-        /// </summary>
-        OutputDefault = OutputToDebug | OutputToConsoleStdOut,
-
-        /// <summary>
-        /// Specifies Explanations are output to all
-        /// </summary>
-        OutputAll = OutputToDebug | OutputToTrace | OutputToConsoleStdOut | OutputToConsoleStdErr,
-
-        /// <summary>
-        /// Show the Thread ID of the Thread evaluating the query (useful in multi-threaded environments)
-        /// </summary>
-        ShowThreadID = 16,
-
-        /// <summary>
-        /// Show the Depth of the Algebra Operator
-        /// </summary>
-        ShowDepth = 32,
-
-        /// <summary>
-        /// Show the Type of the Algebra Operator
-        /// </summary>
-        ShowOperator = 64,
-
-        /// <summary>
-        /// Show the Action being performed (makes it clear whether the explanation marks the start/end of an operation)
-        /// </summary>
-        ShowAction = 128,
-
-        /// <summary>
-        /// Shows Timings for the Query
-        /// </summary>
-        ShowTimings = 256,
-
-        /// <summary>
-        /// Show Intermediate Result Counts at each stage of evaluation
-        /// </summary>
-        ShowIntermediateResultCount = 512,
-
-        /// <summary>
-        /// Shows Basic Information (Depth, Operator and Action)
-        /// </summary>
-        ShowBasic = ShowDepth | ShowOperator | ShowAction,
-
-        /// <summary>
-        /// Shows Default Information (Thread ID, Depth, Operator and Action)
-        /// </summary>
-        ShowDefault = ShowThreadID | ShowDepth | ShowOperator | ShowAction,
-
-        /// <summary>
-        /// Shows All Information
-        /// </summary>
-        ShowAll = ShowThreadID | ShowDepth | ShowOperator | ShowAction | ShowTimings | ShowIntermediateResultCount,
-
-        /// <summary>
-        /// Shows an analysis of BGPs prior to evaluating them
-        /// </summary>
-        /// <remarks>
-        /// This lets you see how many joins, cross products, filters, assignments etc must be applied in each BGP
-        /// </remarks>
-        AnalyseBgps = 1024,
-
-        /// <summary>
-        /// Shows an analysis of Joins prior to evaluating them
-        /// </summary>
-        /// <remarks>
-        /// This lets you see whether the join is a join/cross product and in the case of a Minus whether the RHS can be ignored completely
-        /// </remarks>
-        AnalyseJoins = 2048,
-
-        /// <summary>
-        /// Shows an analysis of Named Graphs used by a Graph clause prior to evaluating them
-        /// </summary>
-        /// <remarks>
-        /// This lets you see how many graphs a given Graph clause will operate over.  As the Graph clause in SPARQL is defined as the union of evaluating the inner operator over each named graph in the dataset graph clauses applied to datasets with many named graphs can be expensive.
-        /// </remarks>
-        AnalyseNamedGraphs = 4096,
-
-        /// <summary>
-        /// Sets whether Evaluation should be simulated (means timings will not be accurate but allows you to explain queries without needing actual data to evaluate them against)
-        /// </summary>
-        Simulate = 8092,
-
-        /// <summary>
-        /// Shows all analysis information
-        /// </summary>
-        AnalyseAll = AnalyseBgps | AnalyseJoins | AnalyseNamedGraphs,
-
-        /// <summary>
-        /// Basic Explanation Level (Console Standard Output and Basic Information)
-        /// </summary>
-        Basic = OutputToConsoleStdOut | ShowBasic,
-
-        /// <summary>
-        /// Default Explanation Level (Default Outputs and Default Information)
-        /// </summary>
-        Default = OutputDefault | ShowDefault,
-
-        /// <summary>
-        /// Detailed Explanation Level (Default Outputs and All Information)
-        /// </summary>
-        Detailed = OutputDefault | ShowAll,
-
-        /// <summary>
-        /// Full Explanation Level (All Outputs, All Information and All Analysis)
-        /// </summary>
-        Full = OutputAll | ShowAll | AnalyseAll,
-
-        /// <summary>
-        /// Basic Explanation Level with Query Evaluation simulated
-        /// </summary>
-        BasicSimulation = Basic | Simulate,
-
-        /// <summary>
-        /// Default Explanation Level with Query Evaluation simulated
-        /// </summary>
-        DefaultSimulation = Default | Simulate,
-
-        /// <summary>
-        /// Detailed Explanation Level with Query Evaluation simulated
-        /// </summary>
-        DetailedSimulation = Detailed | Simulate,
-
-        /// <summary>
-        /// Full Explanation Level with Query Evaluation simulated
-        /// </summary>
-        FullSimulation = Full | Simulate,
-    }
-
-    /// <summary>
     /// A Query Processor which evaluates queries while printing explanations to any/all of Debug, Trace, Console Standard Output and Console Standard Error.
     /// </summary>
     public class ExplainQueryProcessor
@@ -466,7 +305,7 @@ namespace VDS.RDF.Query
                         {
                             // Query is over entire dataset/default Graph since no named Graphs are explicitly specified
                             PrintExplanations("Graph clause uses variable ?" + gvar + " which accesses all named graphs provided by the dataset");
-                            activeGraphs.AddRange(context.Data.GraphUris.Select(u => u.ToSafeString()));
+                            activeGraphs.AddRange(context.Data.GraphNames.Select(u => u.ToSafeString()));
                         }
                     }
 
@@ -938,45 +777,6 @@ namespace VDS.RDF.Query
         public override BaseMultiset ProcessZeroOrMorePath(ZeroOrMorePath path, SparqlEvaluationContext context)
         {
             return ExplainAndEvaluate<ZeroOrMorePath>(path, context, base.ProcessZeroOrMorePath);
-        }
-    }
-
-    class ExplainDataset 
-        : WrapperDataset
-    {
-        public ExplainDataset(ISparqlDataset dataset) 
-            : base(dataset) {}
-
-        public ExplainQueryProcessor Processor { get; set; }
-
-        public override void SetActiveGraph(Uri graphUri)
-        {
-            if (Processor != null)
-            {
-                if (Processor.HasFlag(ExplanationLevel.AnalyseNamedGraphs))
-                {
-                    Processor.PrintExplanations("Switching to named graph " + graphUri.ToSafeString());
-                }
-            }
-            base.SetActiveGraph(graphUri);
-        }
-
-        public override void SetActiveGraph(IEnumerable<Uri> graphUris)
-        {
-            IList<Uri> gs = graphUris as IList<Uri> ?? graphUris.ToList();
-            if (Processor != null)
-            {
-                if (Processor.HasFlag(ExplanationLevel.AnalyseNamedGraphs))
-                {
-                    Processor.PrintExplanations("Switching to named graph as merge of the following graphs:");
-                    foreach (Uri graphUri in gs)
-                    {
-                        Processor.PrintExplanations(graphUri.ToSafeString());
-                        
-                    }
-                }
-            }
-            base.SetActiveGraph(gs);
         }
     }
 }
